requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bitly-api-py3==0.3
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
django-bitly==1.2.1
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
